+++ src/chemex/experiments/catalog/cpmg_ch3_1h_dq.py
@@ -55,7 +55,7 @@ class CpmgCh31HDqSettings(CpmgSettings):
             List of initial state terms for the Liouvillian calculation.
 
         """
-        return [f"2ixsz{self.suffix_start}"]
+        return self.get_start_terms("2ixsz")
 
     @computed_field
     @property
@@ -66,7 +66,7 @@ class CpmgCh31HDqSettings(CpmgSettings):
             Detection term for the Liouvillian calculation.
 
         """
-        return f"[2ixsz{self.suffix_detect}]"
+        return self.get_detection_expression("[2ixsz]")
 
 
 class CpmgCh31HDqConfig(
@@ -78,7 +78,7 @@ class CpmgCh31HDqConfig(
 ):
     @property
     def to_be_fitted(self) -> ToBeFitted:
-        state = self.experiment.observed_state
+        state = self.experiment.primary_state
 
         to_be_fitted = ToBeFitted(rates=[f"r2_i_{state}"], model_free=[f"tauc_{state}"])