+++ src/chemex/experiments/catalog/relaxation_hznz.py
@@ -37,7 +37,7 @@ class RelaxationHzNzSettings(RelaxationSettings):
             List of initial state terms for the Liouvillian calculation.
 
         """
-        return [f"2izsz{self.suffix_start}"]
+        return self.get_start_terms("2izsz")
 
     @computed_field
     @property
@@ -48,7 +48,7 @@ class RelaxationHzNzSettings(RelaxationSettings):
             Detection term for the Liouvillian calculation.
 
         """
-        return f"[2izsz{self.suffix_detect}]"
+        return self.get_detection_expression("[2izsz]")
 
 
 class RelaxationHzNzConfig(
@@ -60,7 +60,7 @@ class RelaxationHzNzConfig(
 ):
     @property
     def to_be_fitted(self) -> ToBeFitted:
-        state = self.experiment.observed_state
+        state = self.experiment.primary_state
         return ToBeFitted(
             rates=[f"r1a_is_{state}"],
             model_free=[f"tauc_{state}", f"s2_{state}", f"khh_{state}"],